Move `printTupleType` into `array.js`
