+++ src/language-js/print/type-annotation.js
@@ -283,26 +283,6 @@ function printFunctionType(path, options, print) {
   return group(parts);
 }
 
-// `TSTupleType` and `TupleTypeAnnotation`
-function printTupleType(path, options, print) {
-  const { node } = path;
-  const typesField = node.type === "TSTupleType" ? "elementTypes" : "types";
-  const types = node[typesField];
-  const isEmptyTuple = types.length === 0;
-  const openBracket = "[";
-  const closeBracket = "]";
-  if (isEmptyTuple) {
-    return printEmptyArray(path, openBracket, closeBracket, options);
-  }
-  return group([
-    openBracket,
-    indent([softline, printArrayItems(path, options, typesField, print)]),
-    ifBreak(shouldPrintComma(options, "all") ? "," : ""),
-    softline,
-    closeBracket,
-  ]);
-}
-
 // `TSIndexedAccessType`, `IndexedAccessType`, and `OptionalIndexedAccessType`
 function printIndexedAccessType(path, options, print) {
   const { node } = path;
